+++ tests/test_doitlive.py
@@ -81,11 +81,23 @@ def test_custom_speed(runner):
     assert '123456789' in result.output
 
 
+def test_bad_theme(runner):
+    result = runner.invoke(cli, ['-p', 'thisisnotatheme'])
+    assert result.exit_code > 0
+
+def test_preview(runner):
+    assert runner.invoke(cli, ['--preview']).exit_code == 0
+    assert runner.invoke(cli, ['-P']).exit_code == 0
+
 class TestTermString:
 
     @pytest.fixture
     def ts(self):
         return TermString('foo')
+
+    @pytest.fixture
+    def ts_blank(self):
+        return TermString('')
 
     def test_str(self, ts):
         assert str(ts) == 'foo'
@@ -105,3 +117,18 @@ class TestTermString:
 
     def test_dim(self, ts):
         assert str(ts.dim) == style('foo', dim=True)
+
+    def test_underlined(self, ts):
+        assert str(ts.underlined) == style('foo', underline=True)
+
+    def test_paren(self, ts, ts_blank):
+        assert str(ts.paren) == '(foo)'
+        assert str(ts_blank.paren) == ''
+
+    def test_square(self, ts, ts_blank):
+        assert str(ts.square) == '[foo]'
+        assert str(ts_blank.square) == ''
+
+    def test_curly(self, ts, ts_blank):
+        assert str(ts.curly) == '{foo}'
+        assert str(ts_blank.curly) == ''