+++ HISTORY.rst
@@ -4,12 +4,15 @@ Changelog
 1.0 (unreleased)
 ******************
 
-- Prompts variables can have ANSI colors and styles.
+- Added themes!
+- Prompt variables can have ANSI colors and styles.
+- ``{hostname}`` can be included in prompt.
+- ``{git_branch}`` can be included in prompt.
 - Prompt variable``{{full_cwd}}`` renamed to ``{{cwd}}``.
 - Prompt variable ``{cwd}`` renamed to ``{dir}``.
 - Short option for `--speed` is now ``-s``.
 - Short option for ``--shell`` is now ``-S``.
-- ``{hostname}`` can be included in prompt.
+- Changed default prompt.
 - ``run`` and ``magictype`` receive prompt_template instead of a prompt function.
 - Remove unnecessary ``PromptState`` class.